Makes backup directory path absolute
Ensures the backup directory path is absolute if a relative
path starting with "./" is provided. This prevents issues
where the backup script might be executed from different
directories, leading to incorrect backup locations.

Also, adds a check to verify that the images backup directory
was successfully created. If not, the script will exit
with an error message.